Fix Solo12BulletRobot. Adding SoloBaseRobot base class
